m_len_lineusers.student_id項目の型不正

+++ database/migrations/2022_09_28_125936_create_m_len_lineusers_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->collation = 'utf8mb4_general_ci';
 
             $table->id();
-            $table->integer('student_id')->unsigned();
+            $table->unsignedBigInteger('student_id')->unsigned();
             $table->string('lineDisplayName',255);
             $table->string('lineUserId',255);
             $table->timestamps();
